[gunicorn] Move _AddressType to file ._types.pyi
