+++ stubs/gunicorn/gunicorn/_types.pyi
@@ -1,5 +1,5 @@
 from collections.abc import Iterable
-from typing import Any, Callable
+from typing import Any, Callable, LiteralString
 
 type _StatusType = str
 type _HeadersType = list[tuple[str, str]]
@@ -9,3 +9,8 @@ type _StartResponseType = Callable[[_StatusType, _HeadersType], None]
 type _ResponseBodyType = Iterable[bytes]
 
 type _WSGIAppType = Callable[[_EnvironType, _StartResponseType], _ResponseBodyType]
+
+type _UnixSocketPathType = str
+type _FileDescriptorType = int
+type _TcpAddressType = tuple[LiteralString, int]
+type _AddressType = _UnixSocketPathType | _FileDescriptorType | _TcpAddressType
